+++ client/src/components/play-lab/booking-form/components/stages/success/sucess.tsx
@@ -2,7 +2,7 @@ import { CheckCircle } from 'lucide-react'
 
 import { Alert, AlertDescription, AlertTitle } from '@ui-components/alert'
 
-import { useFormStage } from '../../zustand/form-stage'
+import { useFormStage } from '../../../zustand/form-stage'
 
 export function Success() {
     const formStage = useFormStage((store) => store.formStage)
