+++ client/src/components/play-lab/booking-form/pages/play-lab-booking-page.tsx
@@ -5,16 +5,16 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { Form } from '@ui-components/form'
 import { Separator } from '@ui-components/separator'
 
-import { BackButton } from '../form/components/back-button'
-import { BookingTypeSelector } from '../form/components/booking-type-selector'
-import { CasualProgramSelector } from '../form/components/casual-program-selector'
-import { CustomerDetails } from '../form/components/customer-details'
-import { Payment } from '../form/components/payment'
-import { Stepper } from '../form/components/stepper'
-import { StudioSelector } from '../form/components/studio-selector'
-import { Success } from '../form/components/sucess'
-import { TermProgramSelector } from '../form/components/term-program-selector'
-import { PlayLabBookingForm, formSchema } from '../form/form-schema'
+import { BackButton } from '../components/common/back-button'
+import { Stepper } from '../components/common/stepper'
+import { PlayLabBookingForm, formSchema } from '../components/form-schema'
+import { CustomerDetails } from '../components/stages/customer-details/customer-details'
+import { Payment } from '../components/stages/payment/payment'
+import { BookingTypeSelector } from '../components/stages/program-selection/booking-type-selector'
+import { CasualProgramSelector } from '../components/stages/program-selection/casual-program-selector'
+import { StudioSelector } from '../components/stages/program-selection/studio-selector'
+import { TermProgramSelector } from '../components/stages/program-selection/term-program-selector'
+import { Success } from '../components/stages/success/sucess'
 
 export function PlayLabBookingPage() {
     const form = useForm<PlayLabBookingForm>({
